# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GUG Apis ",
        "contact": {
            "name": "Yosof Bayan",
            "url": "https://wa.me/+963967213544",
            "email": "yosofbayan75@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "local Base URL"
        },
        {
            "url": "http://gug-back.icrcompany.net/api/v1",
            "description": "Develop Base URL"
        }
    ],
    "paths": {
...
        "/admin/sections": {
            "get": {
                "tags": [
                    "Admin",
                    "Admin - Section"
                ],
                "summary": "get services sections data",
                "operationId": "admin/super-sections",
                "parameters": [
                    {
                        "name": "trash",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ],
                            "example": 1
                        }
                    },
                    {
                        "name": "paginate",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin",
                    "Admin - Section"
                ],
                "summary": "Store Super Section data",
                "description": "Store Super Section with the provided information",
                "operationId": "post-super-section",
                "requestBody": {
                    "description": "Section data",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "logo"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Arabic section name "
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "English section name "
                                    },
                                    "logo": {
                                        "description": "Image file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "images[0]": {
                                        "description": "Image file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Section stored successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Section stored successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/sections/{id}": {
            "post": {
                "tags": [
                    "Admin",
                    "Admin - Section"
                ],
                "summary": "Update Section data",
                "description": "Update Section with the provided information",
                "operationId": "store-section",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "section id to update ",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "_method",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "PUT",
                            "enum": [
                                "PUT"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "description": "Section data",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Arabic section name "
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "English section name "
                                    },
                                    "logo": {
                                        "description": "Image file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "images[0]": {
                                        "description": "Image file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "delete_images[0]": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Section updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Section udpated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin",
                    "Admin - Section"
                ],
                "summary": "Delete an section or brand",
                "operationId": "139e52fec94f3be1f5e9879fd97d3c2d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the section or brand to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/admin/sections/{section_id}": {
            "get": {
                "tags": [
                    "Admin",
                    "Admin - Section"
                ],
                "summary": "get section data ",
                "operationId": "admin/section",
                "parameters": [
                    {
                        "name": "section_id",
                        "in": "path",
                        "description": "pass the section ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}